pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "autodos"
version = "1.0.0"
description = "Automated Denial-of-Service Attack for LLMs"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}

dependencies = [
    "pydantic>=2.0.0",
    "pyyaml>=6.0",
    "json-repair>=0.25.0",
]

[tool.setuptools]
packages = ["autodos"]
